+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## [1.1.1]
+### Fixed
+- fixed consignee postcode for Latvia
+- fixed Order::getByCartId incompatibility for Prestashop 1.6
+
+## [1.1.0] - 2021-11-23
+### Init
+- release of the completed module
+
 ## [0.1.0] - 2021-08-13
 ### Added
 - created working main functions